fix: don't delete resolv.conf if it is a symlink

+++ xtask/buildenv/scripts/customize.sh
@@ -4,4 +4,9 @@ root_dir=$1;
 
 # We remove those files as they are not needed and would cause the build to be non-deterministic.
 rm "${root_dir}/etc/hostname"
-rm "${root_dir}/etc/resolv.conf"
+
+# Avoid deleting resolv.conf if it is symlink as this is an indication that another
+# service like systemd-resolved is managing it
+if [ ! -L "${root_dir}/etc/resolv.conf" ]; then
+    rm "${root_dir}/etc/resolv.conf"
+fi
